backend/requirements.txt
# Core
Django==5.1.6
djangorestframework==3.15.2
django-cors-headers==4.4.0
django-filter==24.3
dj-database-url==2.3.0
psycopg2-binary==2.9.10
whitenoise==6.8.2
gunicorn==23.0.0

# Auth
djangorestframework-simplejwt==5.3.1

# Tasks
celery==5.4.0
redis==5.2.1
django-celery-beat==2.7.0
django-celery-results==2.5.1

# Storage (MinIO / S3-compatible)
boto3==1.35.98
django-storages==1.14.4

# Excel / Data
pandas==2.2.3
openpyxl==3.1.5
xlrd==2.0.1

# PDF
weasyprint==62.3
Pillow==11.1.0
matplotlib==3.10.0

# MinIO client
minio==7.2.11

# Utilities
python-decouple==3.8

# Testing
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==6.0.0
pytest-mock==3.14.0
factory-boy==3.3.1
faker==33.1.0
